SetETGraphicPointers: always update walk animation
This fixes falling back into the north side of a pit when holding the
stick in any direction.

When levitating out of a pit, ET's neck will be raised and falling into
the pit is disabled until ET no longer collides with the pit.
When ET no longer overlaps with the pit, movement is disabled
while the neck lowering animation plays. Each step of the animation
selects a shorter sprite. Since drawing starts from the top, `etVertPos`
and `etHeight` are correspondingly lowered when the sprite is changed.

On the last step, `etVertPos` and `etHeight` are set to ET's normal
height, and it is expected that `.setNormalETGraphicPointers` will
resume drawing the shorter sprite. This works if the stick is not
pressed at the end of the neck animation. `.setToETRestSpritePtrs` will
update `etGraphicPointers0/1` to a sprite of the correct height.

If the stick is pressed, `.determineETAnimationIndex` will advance the
walking/running animation, but this only happens at specific multiples
of `frameCount`.

Setting the sprite pointer to the current animation frame is also
skipped, so ET is drawn with the stale, taller, neck-extended sprite,
but at the lower location. ET's feet now overlap with the pit after
collision has been re-enabled.

This is particularly confusing because occasionally the bug doesn't
trigger if you happen to exit on the right frame.

Fix by always setting the walk animation sprite pointers regardless of
whether the animation frame has advanced.